Create caesar cypher
I wrote a program that asks for a string and an integer to encrypt the
string.

First I created a hash with all caps letters and another one for all
small letters and mapped each letter to a number e.g. hash1: A=>0, B=>1
hash2: a=>0,...

Then I created two more hashes to map the letters shifted by the integer
value, e.g. integer = 5 -> shifted1: 5=>A,6=>B, shifted2: 5=>a,...

Then i used split("") on the string to get an array of single signs and
then print the the letter from shifted according to the letters hash
e.g. string "A", integer 5: letters A=>0, shifted 0=>F