Create substring check
I created a method that takes two parameters, a string and an array of
strings, e.g. called dictionary.
First, an empty Hash named result with default values of 0 is created.
The string is then split into an array and each of the values in the
array is downcased and check whether it includes one of the strings
from the dictionary.
If the check is true, result[word_from_dictionary] +=1