Create stock picker
I wrote a method that takes an array of numbers as an argument and
prints the best day to buy and best day to sell, as the buy day has
to be before the sell day.
The index of the array represents the day and the value the buy or sell
value.
I used a while loop to iterate over the array from index 0 to -2, as
there has to be at least one day to sell after buying.
Then I set up a variable representing the sell day, which has to be at
least buyday+1 and a variable profit which is 0.
If the sell index is not higher than the last index in the array and if
the sell value - buy value is bigger than the profit, the result is
stored in profit and buy index is stored in best_buy and sell index is
stored in best_sell.
Whether or not the profit is higher, the sell index is incremented by 1
Once sell index = array.length-1, buy is incremented and sell is buy+1